+++ lib/diffo_example/application.ex
@@ -9,6 +9,11 @@ defmodule DiffoExample.Application do
 
   @impl true
   def start(_type, _args) do
-    Supervisor.start_link([], strategy: :one_for_one)
+    children = [
+      {Plug.Cowboy, scheme: :http, plug: DiffoExample.Nbn.Router, options: [port: 4000]},
+      {Task, &DiffoExample.Nbn.Initializer.init/0}
+    ]
+
+    Supervisor.start_link(children, strategy: :one_for_one, name: DiffoExample.Supervisor)
   end
 end