+++ lib/nbn/resources/nni.ex
@@ -22,7 +22,12 @@ defmodule DiffoExample.Nbn.Nni do
 
   use Ash.Resource,
     fragments: [BaseInstance],
-    domain: Nbn
+    domain: Nbn,
+    extensions: [AshJsonApi.Resource]
+
+  json_api do
+    type "nni"
+  end
 
   resource do
     description "An Ash Resource representing a Network-to-Network Interface (NNI)"