+++ lib/nbn/resources/nni_group.ex
@@ -24,7 +24,12 @@ defmodule DiffoExample.Nbn.NniGroup do
 
   use Ash.Resource,
     fragments: [BaseInstance],
-    domain: Nbn
+    domain: Nbn,
+    extensions: [AshJsonApi.Resource]
+
+  json_api do
+    type "nniGroup"
+  end
 
   resource do
     description "An Ash Resource representing an NNI Group"